requirements.txt
tensorflow
attrs
